% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NAIC.R
\docType{data}
\name{comauto}
\alias{comauto}
\title{Complete triangle of losses for many commercial auto companies. Data is taken from the CAS.}
\format{Data frame with 400 observations of 14 variables}
\source{
\url{http://www.casact.org/research/reserve_data/ppauto_pos.csv}
}
\description{
Complete triangle of losses for many commercial auto companies. Data is taken from the CAS.
}

