% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NAIC.R
\name{dfWC}
\alias{dfWC}
\title{Complete triangle of losses for 4 workers compensation companies. Data is taken from the CAS.}
\format{Data frame with 400 observations of 14 variables}
\source{
\url{http://www.casact.org/research/reserve_data/wkcomp_pos.csv}
}
\description{
Complete triangle of losses for 4 workers compensation companies. Data is taken from the CAS.
}

