% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InstallSuggestions.R
\name{InstallSuggestions}
\alias{InstallSuggestions}
\title{InstallSuggestions}
\usage{
InstallSuggestions(...)
}
\arguments{
\item{...}{Arguments to be passed to install.packages.}
}
\value{
A character vector indicating which packages were installled.
}
\description{
This function will install the full set of suggested packages.
}
\examples{

\dontrun{
  InstallSuggestions()
}

}

