% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NAIC.R
\docType{data}
\name{MultiTri}
\alias{MultiTri}
\title{MultiTri}
\source{
\url{http://www.casact.org/research/reserve_data/ppauto_pos.csv}

\url{http://www.casact.org/research/reserve_data/wkcomp_pos.csv}

\url{http://www.casact.org/research/reserve_data/othliab_pos.csv}
}
\description{
Complete triangle of losses for the same 4 private passenger auto and workers comp companies. Data is taken from the CAS.

\itemize{
   \item GroupCode: NAIC company code
   \item Company 
   \item AccidentYear
   \item DevelopmentYear
   \item Lag: Integer indicating the development lag in years.
   \item CumulativeIncurred
   \item CumulativePaid
   \item IBNR: Bulk and IBNR reserves
   \item DirectEP: Direct and assumed earned premium, gross of reinsurance.
   \item CededEP: Earned premium ceded to reinsurers.
   \item NetEP: Premium earned net of reinsurance.
   \item Single: Either 1 or 0. 1 indicates a single entity, 0 indicates a group entity.
   \item Reserve1997: Posted reserves as of year-end 1997.
}
}
\seealso{
\code{\link{NAIC}}
}
