% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{NFL}
\alias{NFL}
\title{NFL game data}
\format{Data frame with 25 variables
   \code{Week}, \code{Day}, \code{Date}, \code{Outcome}
   , \code{OT}, \code{Record}, \code{Home}, \code{Opponent}
   , \code{ThisTeamScore}, \code{OpponentScore}
   , \code{ThisTeam1D}, \code{ThisTeamTotalYards}
   , \code{ThisTeamPassYards}, \code{ThisTeamRushYards}
   , \code{ThisTeamTO}, \code{Opponent1D}
   , \code{OpponentTotalYards}, \code{OpponentPassYards}
   , \code{OpponentRushYards}, \code{OpponentTO}
   , \code{GameDate}, \code{TODiff}
   , \code{Win}, \code{ThisTeam}, \code{ScoreDifference}}
\usage{
NFL
}
\description{
Results of over 13,000 NFL games
}
\keyword{datasets}

