% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InstallSuggestions.R
\name{PackageDependencies}
\alias{PackageDependencies}
\title{PackageDependencies}
\usage{
PackageDependencies(x, type = c("Depends", "Imports", "Suggests", "Enhances",
  "LinkingTo"))
}
\arguments{
\item{x}{The name of the packages}

\item{type}{What sort of dependency are you looking for? Permissible values are "Depends", "Imports", 
"Suggests", "Enhances" and "LinkingTo"}
}
\description{
PackageDependencies will examine the DESCRIPTION file of a package and return a vector of 
             dependencies.
}
\examples{

\dontrun{
  PackageDependencies("ggplot2", "Depends")
}

}
