% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InstallSuggestions.R
\name{PackageIsInstalled}
\alias{PackageIsInstalled}
\title{PackageIsInstalled}
\usage{
PackageIsInstalled(pkgs)
}
\arguments{
\item{pkgs}{Character vector of package names.}
}
\value{
A logical vector indicating whether the packages have been installed.
}
\description{
This function will indicate whether a package has been installed. It uses the installed.packages 
function for this.
}
\examples{

\dontrun{
  PackageIsInstalled("ggplot2")
}

}

