% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InstallSuggestions.R
\name{PackageNeedsUpdating}
\alias{PackageNeedsUpdating}
\title{PackageNeedsUpdating}
\usage{
PackageNeedsUpdating(pkgs)
}
\arguments{
\item{pkgs}{Character vector of package names.}
}
\value{
A logical vector indicating whether the packages need to be updated.
}
\description{
This function will indicate whether a package needs to be updated. It uses the old.packages 
function for this.
}
\examples{

\dontrun{
  PackageNeedsUpdating("ggplot2")
}

}

