% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raw.R
\docType{package}
\name{raw}
\alias{raw}
\alias{raw-package}
\title{raw - R Actuarial Workshops}
\description{
This is a packge which stores data used in R workshops sponsored by the Casualty Actuarial Society.
}
\details{
For more information, read the package vignette. To access the package vignette, type in the code shown in the 
Examples section (the line without the hash marks).
}
\examples{
\dontrun{
vignette(topic = "raw")
}

}

