% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{StateExperience}
\alias{StateExperience}
\title{State data}
\format{A data frame with 7 variables:
   \code{Fullname}, \code{PolicyYear}, \code{Postal}
   , \code{NumPolicies}, \code{PolicyGrowth}, \code{Lambda}
   , \code{BasePolicies}, \code{NumClaims}, \code{Region}}
\usage{
StateExperience
}
\description{
Random illustrative data
}
\keyword{datasets}

