% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/COTOR.R
\docType{data}
\name{COTOR2}
\alias{COTOR2}
\alias{COTOR}
\alias{COTOR1}
\alias{COTOR3}
\alias{COTOR4}
\alias{COTOR5}
\title{Committee on the Theory of Risk (COTOR)}
\usage{
data(COTOR2)
}
\description{
The \href{http://www.casact.org/research/cotor/}{Casualty Actuarial Society's Committee on Theory of Risk (COTOR)} 
issued several challenges as part of their work. These involved artificial constructed loss data. Actuaries
were invited to construct models to estimate appropriate premiums for risk transfer.

Note that there is no COTOR1.

\strong{COTOR2}

COTOR2 is a vector containing 250 observations of claim values.

\href{http://www.casact.org/research/cotor/index.cfm?fa=round2}{COTOR round 2}

\strong{COTOR3}

COTOR3 is a data frame with 490 observations of 2 variables. The values were generated by the great Stuart Klugman.

\itemize{
   \item Year: an integer indicating the year in which the claim occurred.
   \item Amount: the amount of the claim.
}

\href{http://www.casact.org/research/cotor/index.cfm?fa=round3}{COTOR round 3}

\strong{COTOR4}

COTOR4 is a data frame with 2500 rows and 2 columns.

\itemize{
   \item Year: an integer indicating the year in which the claim occurred.
   \item Amount: the amount of the claim.
}

\href{http://www.casact.org/research/cotor/index.cfm?fa=round4}{COTOR round 4}

\strong{COTOR5}

COTOR5 is a data frame with 4849 rows and 5 columns

\itemize{
   \item Risk: A single letter indicating the individual risk to which the claim belongs.
   \item Year: an integer indicating the year in which the claim occurred.
   \item MSA: simulated metropolitan statistical area.
   \item Amount: the amount of the claim.
}

\href{http://www.casact.org/research/cotor/index.cfm?fa=round5}{COTOR round 5}
}
