% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Hurricane}
\alias{Hurricane}
\title{Hurricane data}
\format{A data frame with 7 variables:
   \code{Year}, \code{Number}, \code{ISO_Time}
   , \code{Latitude}, \code{Longitude}, \code{Wind}}
\usage{
Hurricane
}
\description{
Listing of the position of and windspeed of hurricanes.

Data is taken from the National Hurricane Center, a division of the
National Oceanic and Atmospheric Association.
}
\keyword{datasets}

