% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InstallSuggestions.R
\name{InstallSuggestions}
\alias{InstallSuggestions}
\title{InstallSuggestions}
\usage{
InstallSuggestions(pkg = "raw", quiet, ...)
}
\arguments{
\item{pkg}{Name of the package for which you want to install suggestions}

\item{quiet}{Include additional messages while installing?}

\item{...}{Additional named arguments passed to `install.packages`}
}
\value{
A character vector indicating which packages were installled.
}
\description{
This function will install the full set of suggested packages.
}
\examples{

\dontrun{
  InstallSuggestions()
}

}

