% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RegionExperience.R
\docType{data}
\name{StateExperience}
\alias{StateExperience}
\title{State data}
\format{An object of class \code{data.frame} with 500 rows and 6 columns.}
\usage{
data(StateExperience)
}
\description{
Random illustrative data

\itemize{
   \item Region: The name of the region to which the state belongs
   \item State: The name of the state
   \item Postal: Two character postal code for the state
   \item PolicyYear: An integer between 2001 and 2010.
   \item NumPolicies: Integer giving the number of policies in that region for that policy year
   \item NumClaims: Integer giving the number of claims in that region for that policy year
}
}
\seealso{
\code{\link{RegionExperience}}
}
\keyword{datasets}

